% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiKey.R
\name{freecurrencyapi_api_key}
\alias{freecurrencyapi_api_key}
\title{Freecurrencyapi API Key}
\usage{
freecurrencyapi_api_key(force = FALSE)
}
\arguments{
\item{force}{If \code{TRUE}, resets the API key & requires the user to provide a new API key. If \code{FALSE} and an API key
already exists, the key will be printed to the console. If no key exists,
you will be required to enter a key.
\code{force}. Defaults to \code{FALSE}.}
}
\value{
Returns the set API key that has been stored as an enviroment variable.
}
\description{
\href{https://freecurrencyapi.com}{freecurrencyapi.com} requires authentication via an API key. For this package, the API key is saved as a
environmental variable. In interactive mode, using \code{freecurrencyapi_api_key}
will require you to enter an API key. Alternatively, you can also use
\code{Sys.setenv(FREECURRENCYAPI_API_KEY = <key>)} to set the API key manually.
}
