% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getApiStatus.R
\name{get_api_status}
\alias{get_api_status}
\title{Get the current status of the API.}
\usage{
get_api_status()
}
\value{
Returns your current quota. Requests to this endpoint do not count against your quota or rate limit.
}
\description{
Get the current status of the API.
}
