% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHistoricalExchangeRates.R
\name{get_historical_exchange_rates}
\alias{get_historical_exchange_rates}
\title{Get historical exchange rates for a specific date for a single or multiple currencies.}
\usage{
get_historical_exchange_rates(date, base_currency = NULL, currencies = NULL)
}
\arguments{
\item{date}{(required) Date to retrieve historical rates from (format: 2021-12-31)}

\item{base_currency}{The base currency for the conversion (e.g., "USD")}

\item{currencies}{A list of comma seperated currency codes which you want to get (EUR,USD,CAD) By default all available currencies will be shown}
}
\value{
Returns exchange rates for a given time range. Generally, we provide data going back to 1999.
}
\description{
Get historical exchange rates for a specific date for a single or multiple currencies.
}
