% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatestExchangeRates.R
\name{get_latest_exchange_rates}
\alias{get_latest_exchange_rates}
\title{Get exchange rate from the Freecurrencyapi.com API}
\usage{
get_latest_exchange_rates(base_currency = NULL)
}
\arguments{
\item{base_currency}{The base currency for the conversion (e.g., "USD")}
}
\value{
Returns the latest currency exchange rates.
}
\description{
Get exchange rate from the Freecurrencyapi.com API
}
