% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportedCurrencies.R
\name{get_supported_currencies}
\alias{get_supported_currencies}
\title{Get the current status of the API.}
\usage{
get_supported_currencies(currencies = NULL)
}
\arguments{
\item{currencies}{(required) A list of comma seperated currency codes which you want to get (EUR,USD,CAD) By default all available currencies will be shown.}
}
\value{
Returns all our supported currencies.
}
\description{
Get the current status of the API.
}
